pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Noah West", email = "westnoahb@gmail.com"},
]
dependencies = [
    "kivy>=2.1.0",
    "autopep8>=2.0.1",
    "kivymd>=1.1.1",
    "pyserial>=3.5",
    "kivy-garden.graph>=0.4.0",
    "plyer>=2.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
